fixed 64-Bit build - don't define ULONG the microsoft way (unsigned long), used uint32_t instead to keep the CCatDefinitions.h - fixed some warnings to ensure 64-Bit capability

+++ module.c
@@ -60,7 +60,7 @@ static int ccat_bar_init(struct ccat_bar *bar, size_t index,
 	bar->len = pci_resource_len(pdev, index);
 	bar->flags = pci_resource_flags(pdev, index);
 	if (!(IORESOURCE_MEM & bar->flags)) {
-		pr_info("bar%d is no mem_region -> abort.\n", index);
+		pr_info("bar%llu is no mem_region -> abort.\n", (uint64_t)index);
 		return -EIO;
 	}
 
@@ -69,16 +69,16 @@ static int ccat_bar_init(struct ccat_bar *bar, size_t index,
 		pr_info("allocate mem_region failed.\n");
 		return -EIO;
 	}
-	pr_debug("bar%d at [%lx,%lx] len=%lu res: %p.\n", index,
+	pr_debug("bar%llu at [%lx,%lx] len=%lu res: %p.\n", (uint64_t)index,
 		 bar->start, bar->end, bar->len, res);
 
 	bar->ioaddr = ioremap(bar->start, bar->len);
 	if (!bar->ioaddr) {
-		pr_info("bar%d ioremap failed.\n", index);
+		pr_info("bar%llu ioremap failed.\n", (uint64_t)index);
 		release_mem_region(bar->start, bar->len);
 		return -EIO;
 	}
-	pr_debug("bar%d I/O mem mapped to %p.\n", index, bar->ioaddr);
+	pr_debug("bar%llu I/O mem mapped to %p.\n", (uint64_t)index, bar->ioaddr);
 	return 0;
 }
 
@@ -120,12 +120,12 @@ int ccat_dma_init(struct ccat_dma *const dma, size_t channel,
 	dma->size = 2 * memSize - PAGE_SIZE;
 	dma->virt = dma_zalloc_coherent(dev, dma->size, &dma->phys, GFP_KERNEL);
 	if (!dma->virt || !dma->phys) {
-		pr_info("init DMA%d memory failed.\n", channel);
+		pr_info("init DMA%llu memory failed.\n", (uint64_t) channel);
 		return -1;
 	}
 
 	if (request_dma(channel, DRV_NAME)) {
-		pr_info("request dma channel %d failed\n", channel);
+		pr_info("request dma channel %llu failed\n", (uint64_t) channel);
 		ccat_dma_free(dma);
 		return -1;
 	}
@@ -135,10 +135,10 @@ int ccat_dma_init(struct ccat_dma *const dma, size_t channel,
 	memcpy_toio(ioaddr + offset, &addr, sizeof(addr));
 	frame = dma->virt + translateAddr - dma->phys;
 	pr_debug
-	    ("DMA%d mem initialized\n virt:         0x%p\n phys:         0x%llx\n translated:   0x%llx\n pci addr:     0x%08x%x\n memTranslate: 0x%x\n size:         %u bytes.\n",
-	     channel, dma->virt, (uint64_t) (dma->phys), addr,
+	    ("DMA%llu mem initialized\n virt:         0x%p\n phys:         0x%llx\n translated:   0x%llx\n pci addr:     0x%08x%x\n memTranslate: 0x%x\n size:         %llu bytes.\n",
+	     (uint64_t) channel, dma->virt, (uint64_t) (dma->phys), addr,
 	     ioread32(ioaddr + offset + 4), ioread32(ioaddr + offset),
-	     memTranslate, dma->size);
+	     memTranslate, (uint64_t) dma->size);
 	return 0;
 }
